perf(vm): a call frame saved a prefix of each bank when the callee only touches a range
A procedure's registers are numbered above its caller's, so the snapshot taken at every call was
copying registers the callee provably cannot reach: fib touches integer registers 6..15 and the
frame copied 0..15 - six of the sixteen belonged to the caller's own frame. Each procedure now
carries the LOW end of its footprint as well as the high one, and the snapshot is a range.

Three things had to be true for that to pay, and two of them were not obvious:

- The low end can only be computed from the callee's WRITE set. What a callee reads needs no
  protection: an unwritten register still holds the caller's value on return, and FramePush only
  copies the banks aside, so the callee reads the same value either way.
- An absent operand lowers to register 0, so crediting Dest for opcodes that have none - a jump, a
  transfer store, a return - pinned every base back to 0 and made the whole thing inert on the
  first measurement. Opcodes now declare where they write, not just which banks they touch.
- With an indexed loop the offset arithmetic cost more than the copies it saved. Int and float
  snapshot with one block move per bank; strings stay element-wise, being refcounted.

A procedure containing a GOSUB keeps base 0: the GOSUB body can sit outside the procedure's
instruction range, so the scan never sees what it writes - survivable for the high end, which the
program-wide ceiling still covers, not for the low one. FRAMERANGE=0 restores the historic prefix
copy exactly.

Measured against the committed binary built from a worktree with the same audio configuration,
machine idle, min of 11 interleaved runs, output identical in every cell:
fib(32) interpreter 1391 -> 1306 (-6.1%), --jit 1407 -> 1337 (-5.0%), combined 734 -> 702 (-4.4%),
--aot 732 -> 705 (-3.7%); spectral-norm, n-body, fannkuch and intrec within +/-0.6%, as expected -
this removes work from the call path and nowhere else. Integer elements copied per call on fib:
16 -> 10.

Nets: regression 500/500 0 OPTDIFF, AOT 0/707, JIT 0/713, combined 0/707, basc 736/2/1 (the DIFF is
the pre-existing m451), FB examples 323/24/95/31/0 with an identical DIFF list.

+++ src/SedaiExecutionContext.pas
@@ -108,9 +108,14 @@ type
     // only has to protect the registers its CALLEE (and everything the callee calls) can touch,
     // which is usually far below the program-wide width - the caller of a two-line SUB should not
     // pay for the biggest procedure in the program. Pushed by FramePush, read by FramePop.
-    FrameWidthInt: array of Integer;
-    FrameWidthFloat: array of Integer;
-    FrameWidthStr: array of Integer;
+    // The saved part of a bank is a RANGE, not a prefix: a procedure's registers are numbered above
+    // its caller's, so saving from 0 copies registers the callee provably cannot reach. Each entry
+    // therefore packs BOTH ends, (width shl 32) or base - deliberately packed rather than kept in
+    // three more arrays, because a separate base stack costs three more pushes and three more pops
+    // of dynamic-array traffic per call, which measured MORE than the copies it saves.
+    FrameWidthInt: array of Int64;
+    FrameWidthFloat: array of Int64;
+    FrameWidthStr: array of Int64;
     FrameWidthTop: Integer;
 
     // --- Record reclamation marks (RAII, V2 frame / M8 block) ---
